refactor: banner 관련 상태를 useBanner Hook으로 분리
